split server into server.h and server.cpp added comments
