+++ Server/server.cpp
@@ -1,43 +1,55 @@
-#include <cstdio>
-#include <netdb.h>
-#include <sys/types.h>
-#include <sys/socket.h>
-#include <netinet/in.h>
-#include <errno.h>
-#include <cstring>
-#include <arpa/inet.h>
-#include <cstdlib>
-#include <unistd.h>
-#include <signal.h>
-#include <sys/wait.h>
-#include <string>
-#include <string.h>
-
-#include "xmlmanip.h"
-
-//defines
-#define SERVER_TCP_PORT 7000	// Default port
-#define BUFLEN	1024				//Buffer length
-
-struct pdata
-{
-  char* plong;
-  char* plat;
-  char* pip;
-  char* ptime;
-  char* pmac;
-};
-
-//prototypes
-
-void readFromClient(int client_socket);
-pdata rawToPData(char* str);
-void sig_handler (int sig);
-Location pDataToLocation (pdata data);
+/*****************************************************************************************************
+**	SOURCE FILE:	server.cpp		source file for the server.exe which listens for client connections
+**									receives packet containing location info, parses the packet and then
+**									generates an XML file to be read by the website and displayed on a map
+**									and in a table.
+**	 	
+**	PROGRAM:	server
+**
+**	FUNCTIONS:
+**		void readFromClient(int client_socket);
+**		pdata rawToPData(char* str);
+**		void sig_handler (int sig);
+**		Location pDataToLocation (pdata data);
+**
+**	DATE: 		March 13, 2015
+**
+**
+**	DESIGNERS: 	Sebastian Pelka 
+**				Sanders Lee
+**
+**	PROGRAMMER: Sebastian Pelka
+**				Sanders Lee
+**
+**	NOTES:
+** 
+*********************************************************************************************************/
+#include "server.h"
 
 int listen_socket, new_socket;
 pdata *packet;
 
+/*********************************************************************
+** Function: main
+**
+** DATE: March 13, 2015
+**
+** DESIGNER:	Sebastian Pelka
+**				Sanders Lee 
+**
+** PROGRAMMER:	Sebastian Pelka
+**				Sanders Lee
+**
+** INTERFACE:	main()
+**
+** RETURNS:	void
+**
+** NOTES:
+** Main entry point of the program. Creates a Listen socket 
+** to listen for connections. When a connection occurs it forks a new
+** process and passes the new accepted socket to the read from client
+** function which handles the client connection.
+***********************************************************************/
 int main()
 {
 	int 	retval, client_len, status;
@@ -114,6 +126,29 @@ int main()
 	return 0;
 }
 
+/*********************************************************************
+** Function: readFromClient(int)
+**
+** DATE: March 13, 2015
+**
+** DESIGNER:	Sebastian Pelka
+**				Filip Gutica
+**
+** PROGRAMMER:	Sebastian Pelka
+**				Filip Gutica
+**
+** INTERFACE:	readFromClient(int)
+**
+** PARAMETERS:
+**			client_socket		- New socket to read from
+**
+** RETURNS:	void
+**
+** NOTES:
+** Continuosly reads the client_socket until it is disconnected. Converts
+** the received raw data into a pData struct which contains longitude,
+** latitude, ip address, time, and mac address. 
+***********************************************************************/
 void readFromClient(int client_socket)
 {
 	int n, bytes_to_read;
@@ -164,9 +199,32 @@ void readFromClient(int client_socket)
 	return;
 }
 
-/*
-Alex's string to char* parser
-*/
+
+/*********************************************************************
+** Function: rawToPData(char*)
+**
+** DATE: March 13, 2015
+**
+** DESIGNER:	Alex Lam
+**
+** PROGRAMMER:	Alex Lam
+**
+** INTERFACE:	rawToPData(char*)
+**
+** PARAMETERS:
+**			str		- char pointer to be converted to a pData struct
+**
+** REVISIONS:
+**			March 13, 2015 		-Modified by Filip to handle a fifth
+**								 element: mad address
+**
+** RETURNS:	pData
+**
+** NOTES:
+** Splits the received char pointer on comma into 5 separate strings:
+** longitude, latitude, ip, time and mac address. Builds the pData struct
+** using the tokens and returns it.
+***********************************************************************/
 pdata rawToPData(char* str)
 {
 	pdata p;
@@ -201,10 +259,31 @@ pdata rawToPData(char* str)
 
 }
 
-/*
-
-*/
-
+/*********************************************************************
+** Function: pDataToLocation(pdata data)
+**
+** DATE: March 13, 2015
+**
+** DESIGNER:	Sebastian Pelka
+**
+** PROGRAMMER:	Sebastian Pelka
+**
+** INTERFACE:	rawToPData(char*)
+**
+** PARAMETERS:
+**			str		- char pointer to be converted to a pData struct
+**
+** REVISIONS:
+**			March 13, 2015 		-Modified by Filip to handle a fifth
+**								 element: mad address
+**
+** RETURNS:	pData
+**
+** NOTES:
+** Splits the received char pointer on comma into 5 separate strings:
+** longitude, latitude, ip, time and mac address. Builds the pData struct
+** using the tokens and returns it.
+***********************************************************************/
 Location pDataToLocation (pdata data)
 {
 	Location new_location;
@@ -224,7 +303,7 @@ Location pDataToLocation (pdata data)
 }
 
 
-/*******************************************************************************************************
+/**************************************************************************
 ** Function: 	sig_handler
 **
 ** DATE:		February 11, 2015
@@ -243,7 +322,7 @@ Location pDataToLocation (pdata data)
 **
 ** NOTES:
 ** Removes the Message Queue on a SIGINT signal
-******************************************************************************************************/
+*****************************************************************************/
 void sig_handler (int sig)
 {
 	int status;
